Processmedia2 next generation formats
Export tracks.json in new json format
Removed soft subs
Videos are AV1 and h265
Images are webp and avif
Static filenames are 11 characters of base62 rather than old hexstrings
output VTT subs

+++ processmedia2/cleanup_media.py
@@ -1,4 +1,5 @@
 import os
+from datetime import datetime, timedelta
 
 from processmedia_libs.meta_overlay import MetaManagerExtended
 
@@ -12,7 +13,8 @@ VERSION = '0.0.0'
 # Main -------------------------------------------------------------------------
 
 
-def cleanup_media(**kwargs):
+def cleanup_media(grace_timedelta=timedelta(), now=datetime.now(), force=False, **kwargs):
+    assert grace_timedelta.days <=0, 'only negative timedeltas'
     meta_manager = MetaManagerExtended(**kwargs)
     meta_manager.load_all()
 
@@ -21,14 +23,23 @@ def cleanup_media(**kwargs):
         for m in meta_manager.meta_items
         for processed_file in m.processed_files.values()
     }
-    unlinked_files = (f for f in meta_manager.processed_files_manager.scan if f.file_no_ext and f.file_no_ext not in all_known_file_hashs)
+    unlinked_files = (
+        f
+        for f in meta_manager.processed_files_manager.scan
+        if f.file_no_ext and f.file_no_ext not in all_known_file_hashs
+    )
 
     count = 0
     for unlinked_file in unlinked_files:
         if kwargs.get('dryrun'):
-            print(unlinked_file.relative)
+            pass
         else:
-            os.remove(unlinked_file.absolute)
+            mtime = datetime.fromtimestamp(unlinked_file.stats.st_mtime)
+            if ((now - mtime) < grace_timedelta) or force:
+                log.debug(f'removing - {unlinked_file.relative}')
+                os.remove(unlinked_file.absolute)
+            else:
+                log.debug(f'keeping grace - {unlinked_file.relative}')
         count += 1
     log.info('Cleaned up - {} files'.format(count))
 
@@ -36,6 +47,7 @@ def cleanup_media(**kwargs):
 
 def additional_arguments(parser):
     parser.add_argument('--dryrun', action='store_true', help='', default=False)
+    parser.add_argument('--grace_timedelta', action='store', type=lambda d: timedelta(days=int(d)), help='Do not delete files newer than -7 days', default=timedelta(days=-7))
 
 
 if __name__ == "__main__":
